requirements/base.txt
#
# Django Core
Django==1.11.4

#
# Connections / Libs / Etc
pytz==2017.2
django-redis==4.8.0
redis>=2.10.5
requests==2.18.4
psycopg2==2.7.3
celery[redis]==4.1.0
boto3==1.4.6
ipython==6.1.0
asana==0.6.5
stripe==1.62.0
python-intercom==3.1.0
elasticsearch>=5.0.0,<6.0.0
elasticsearch-dsl>=5.0.0,<6.0.0
sparkpost>=1.3.5


#
# Django Contrib
django-environ==0.4.3
django-annoying==0.10.3
django-braces==1.11.0
django-cors-headers==2.1.0
django-filter==1.0.4
django-localflavor==1.5.2
django-model-utils==3.0.0
django-clear-cache==0.3
whitenoise==3.3.0
django-organizations==0.9.2
django-extensions==1.8.1
django-crum==0.7.2
django-material==1.0.0
django-address==0.1.5
django-health-check==2.3.0
awesome-slugify==1.6.5
django-allauth==0.32.0
argon2-cffi==16.3.0
django-fsm==2.6.0
django-cacheops==3.2.1
stream_django==1.3.3
django-rest-auth==0.9.1
django-templated-email==2.2.0

#
# DRF
djangorestframework==3.6.3
coreapi==2.3.1
drfdocs==0.0.11
djangorestframework-jwt==1.11.0
djangorestframework-jsonapi==2.2.0
dry-rest-permissions==0.1.9
django-rest-elasticsearch==0.3.2
django-elasticsearch-dsl==0.2.0
django-rest-swagger==2.1.2
thorn==1.5.0
